# server/ui/api/v1/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "OpenAPI 2.0 documentation for [MailSandbox](https://github.com/axllent/mailsandbox).",
    "title": "MailSandbox API",
    "contact": {
      "name": "GitHub",
      "url": "https://github.com/axllent/mailsandbox"
    },
    "license": {
      "name": "MIT license",
      "url": "https://github.com/axllent/mailsandbox/blob/develop/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/message/{ID}/html-check": {
      "get": {
        "description": "Returns the summary of the message HTML checker.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "other"
        ],
        "summary": "HTML check",
        "operationId": "HTMLCheckParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "HTMLCheckResponse",
            "schema": {
              "$ref": "#/definitions/HTMLCheckResponse"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
    "/api/v1/message/{ID}/link-check": {
      "get": {
        "description": "Returns the summary of the message Link checker.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "other"
        ],
        "summary": "Link check",
        "operationId": "LinkCheckParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "default": "false",
            "x-go-name": "Follow",
            "description": "Follow redirects",
            "name": "follow",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "LinkCheckResponse",
            "schema": {
              "$ref": "#/definitions/LinkCheckResponse"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
...
    "/api/v1/message/{ID}/sa-check": {
      "get": {
        "description": "Returns the SpamAssassin summary (if enabled) of the message.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "other"
        ],
        "summary": "SpamAssassin check",
        "operationId": "SpamAssassinCheckParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "SpamAssassinResponse",
            "schema": {
              "$ref": "#/definitions/SpamAssassinResponse"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
...
    "/api/v1/send": {
      "post": {
        "description": "Send a message via the HTTP API.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "message"
        ],
        "summary": "Send a message",
        "operationId": "SendMessageParams",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "From"
              ],
              "properties": {
                "Attachments": {
                  "description": "Attachments",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "Content",
                      "Filename"
                    ],
                    "properties": {
                      "Content": {
                        "description": "Base64-encoded string of the file content",
                        "type": "string",
                        "example": "iVBORw0KGgoAAAANSUhEUgAAAEEAAAA8CAMAAAAOlSdoAAAACXBIWXMAAAHrAAAB6wGM2bZBAAAAS1BMVEVHcEwRfnUkZ2gAt4UsSF8At4UtSV4At4YsSV4At4YsSV8At4YsSV4At4YsSV4sSV4At4YsSV4At4YtSV4At4YsSV4At4YtSV8At4YsUWYNAAAAGHRSTlMAAwoXGiktRE5dbnd7kpOlr7zJ0d3h8PD8PCSRAAACWUlEQVR42pXT4ZaqIBSG4W9rhqQYocG+/ys9Y0Z0Br+x3j8zaxUPewFh65K+7yrIMeIY4MT3wPfEJCidKXEMnLaVkxDiELiMz4WEOAZSFghxBIypCOlKiAMgXfIqTnBgSm8CIQ6BImxEUxEckClVQiHGj4Ba4AQHikAIClwTE9KtIghAhUJwoLkmLnCiAHJLRKgIMsEtVUKbBUIwoAg2C4QgQBE6l4VCnApBgSKYLLApCnCa0+96AEMW2BQcmC+Pr3nfp7o5Exy49gIADcIqUELGfeA+bp93LmAJp8QJoEcN3C7NY3sbVANixMyI0nku20/n5/ZRf3KI2k6JEDWQtxcbdGuAqu3TAXG+/799Oyyas1B1MnMiA+XyxHp9q0PUKGPiRAau1fZbLRZV09wZcT8/gHk8QQAxXn8VgaDqcUmU6O/r28nbVwXAqca2mRNtPAF5+zoP2MeN9Fy4NgC6RfcbgE7XITBRYTtOE3U3C2DVff7pk+PkUxgAbvtnPXJaD6DxulMLwOhPS/M3MQkgg1ZFrIXnmfaZoOfpKiFgzeZD/WuKqQEGrfJYkyWf6vlG3xUgTuscnkNkQsb599q124kdpMUjCa/XARHs1gZymVtGt3wLkiFv8rUgTxitYCex5EVGec0Y9VmoDTFBSQte2TfXGXlf7hbdaUM9Sk7fisEN9qfBBTK+FZcvM9fQSdkl2vj4W2oX/bRogO3XasiNH7R0eW7fgRM834ImTg+Lg6BEnx4vz81rhr+MYPBBQg1v8GndEOrthxaCTxNAOut8WKLGZQl+MPz88Q9tAO/hVuSeqQAAAABJRU5ErkJggg=="
                      },
                      "ContentID": {
                        "description": "Optional Content-ID (`cid`) for attachment.\nIf this field is set then the file is attached inline.",
                        "type": "string",
                        "example": "mailsandbox-logo"
                      },
                      "ContentType": {
                        "description": "Optional Content Type for the the attachment.\nIf this field is not set (or empty) then the content type is automatically detected.",
                        "type": "string",
                        "example": "image/png"
                      },
                      "Filename": {
                        "description": "Filename",
                        "type": "string",
                        "example": "mailsandbox.png"
                      }
                    }
                  }
                },
                "Bcc": {
                  "description": "Bcc recipients email addresses only",
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "jack@example.com"
                  ]
                },
                "Cc": {
                  "description": "Cc recipients",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "Email"
                    ],
                    "properties": {
                      "Email": {
                        "description": "Email address",
                        "type": "string",
                        "example": "manager@example.com"
                      },
                      "Name": {
                        "description": "Optional name",
                        "type": "string",
                        "example": "Manager"
                      }
                    }
                  }
                },
                "From": {
                  "description": "\"From\" recipient",
                  "type": "object",
                  "required": [
                    "Email"
                  ],
                  "properties": {
                    "Email": {
                      "description": "Email address",
                      "type": "string",
                      "example": "john@example.com"
                    },
                    "Name": {
                      "description": "Optional name",
                      "type": "string",
                      "example": "John Doe"
                    }
                  }
                },
                "HTML": {
                  "description": "Message body (HTML)",
                  "type": "string",
                  "example": "\u003cdiv style=\"text-align:center\"\u003e\u003cp style=\"font-family: arial; font-size: 24px;\"\u003eMailSandbox is \u003cb\u003eawesome\u003c/b\u003e!\u003c/p\u003e\u003cp\u003e\u003cimg src=\"cid:mailsandbox-logo\" /\u003e\u003c/p\u003e\u003c/div\u003e"
                },
                "Headers": {
                  "description": "Optional headers in {\"key\":\"value\"} format",
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "example": {
                    "X-IP": "1.2.3.4"
                  }
                },
                "ReplyTo": {
                  "description": "Optional Reply-To recipients",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "Email"
                    ],
                    "properties": {
                      "Email": {
                        "description": "Email address",
                        "type": "string",
                        "example": "secretary@example.com"
                      },
                      "Name": {
                        "description": "Optional name",
                        "type": "string",
                        "example": "Secretary"
                      }
                    }
                  }
                },
                "Subject": {
                  "description": "Subject",
                  "type": "string",
                  "example": "MailSandbox message via the HTTP API"
                },
                "Tags": {
                  "description": "MailSandbox tags",
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "Tag 1",
                    "Tag 2"
                  ]
                },
                "Text": {
                  "description": "Message body (text)",
                  "type": "string",
                  "example": "MailSandbox is awesome!"
                },
                "To": {
                  "description": "\"To\" recipients",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "Email"
                    ],
                    "properties": {
                      "Email": {
                        "description": "Email address",
                        "type": "string",
                        "example": "jane@example.com"
                      },
                      "Name": {
                        "description": "Optional name",
                        "type": "string",
                        "example": "Jane Doe"
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SendMessageResponse"
          },
          "400": {
            "$ref": "#/responses/JSONErrorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "HTMLCheckResponse": {
      "description": "Response represents the HTML check response struct",
      "type": "object",
      "properties": {
        "Platforms": {
          "description": "All platforms tested, mainly for the web UI",
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "Total": {
          "$ref": "#/definitions/HTMLCheckTotal"
        },
        "Warnings": {
          "description": "List of warnings from tests",
          "type": "array",
          "items": {
            "$ref": "#/definitions/HTMLCheckWarning"
          }
        }
      },
      "x-go-name": "Response",
      "x-go-package": "github.com/axllent/mailsandbox/internal/htmlcheck"
    },
    "HTMLCheckResult": {
      "description": "Result struct",
      "type": "object",
      "properties": {
        "Family": {
          "description": "Family eg: Outlook, Mozilla Thunderbird",
          "type": "string"
        },
        "Name": {
          "description": "Friendly name of result, combining family, platform \u0026 version",
          "type": "string"
        },
        "NoteNumber": {
          "description": "Note number for partially supported if applicable",
          "type": "string"
        },
        "Platform": {
          "description": "Platform eg: ios, android, windows",
          "type": "string"
        },
        "Support": {
          "description": "Support [yes, no, partial]",
          "type": "string"
        },
        "Version": {
          "description": "Family version eg: 4.7.1, 2019-10, 10.3",
          "type": "string"
        }
      },
      "x-go-name": "Result",
      "x-go-package": "github.com/axllent/mailsandbox/internal/htmlcheck"
    },
    "HTMLCheckScore": {
      "description": "Score struct",
      "type": "object",
      "properties": {
        "Found": {
          "description": "Number of matches in the document",
          "type": "integer",
          "format": "int64"
        },
        "Partial": {
          "description": "Total percentage partially supported",
          "type": "number",
          "format": "float"
        },
        "Supported": {
          "description": "Total percentage supported",
          "type": "number",
          "format": "float"
        },
        "Unsupported": {
          "description": "Total percentage unsupported",
          "type": "number",
          "format": "float"
        }
      },
      "x-go-name": "Score",
      "x-go-package": "github.com/axllent/mailsandbox/internal/htmlcheck"
    },
    "HTMLCheckTotal": {
      "description": "Total weighted result for all scores",
      "type": "object",
      "properties": {
        "Nodes": {
          "description": "Total number of HTML nodes detected in message",
          "type": "integer",
          "format": "int64"
        },
        "Partial": {
          "description": "Overall percentage partially supported",
          "type": "number",
          "format": "float"
        },
        "Supported": {
          "description": "Overall percentage supported",
          "type": "number",
          "format": "float"
        },
        "Tests": {
          "description": "Total number of tests done",
          "type": "integer",
          "format": "int64"
        },
        "Unsupported": {
          "description": "Overall percentage unsupported",
          "type": "number",
          "format": "float"
        }
      },
      "x-go-name": "Total",
      "x-go-package": "github.com/axllent/mailsandbox/internal/htmlcheck"
    },
    "HTMLCheckWarning": {
      "description": "Warning represents a failed test",
      "type": "object",
      "properties": {
        "Category": {
          "description": "Category [css, html]",
          "type": "string"
        },
        "Description": {
          "description": "Description",
          "type": "string"
        },
        "Keywords": {
          "description": "Keywords",
          "type": "string"
        },
        "NotesByNumber": {
          "description": "Notes based on results",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "Results": {
          "description": "Test results",
          "type": "array",
          "items": {
            "$ref": "#/definitions/HTMLCheckResult"
          }
        },
        "Score": {
          "$ref": "#/definitions/HTMLCheckScore"
        },
        "Slug": {
          "description": "Slug identifier",
          "type": "string"
        },
        "Tags": {
          "description": "Tags",
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "Title": {
          "description": "Friendly title",
          "type": "string"
        },
        "URL": {
          "description": "URL to caniemail.com",
          "type": "string"
        }
      },
      "x-go-name": "Warning",
      "x-go-package": "github.com/axllent/mailsandbox/internal/htmlcheck"
    },
    "Link": {
      "description": "Link struct",
      "type": "object",
      "properties": {
        "Status": {
          "description": "HTTP status definition",
          "type": "string"
        },
        "StatusCode": {
          "description": "HTTP status code",
          "type": "integer",
          "format": "int64"
        },
        "URL": {
          "description": "Link URL",
          "type": "string"
        }
      },
      "x-go-package": "github.com/axllent/mailsandbox/internal/linkcheck"
    },
    "LinkCheckResponse": {
      "description": "Response represents the Link check response",
      "type": "object",
      "properties": {
        "Errors": {
          "description": "Total number of errors",
          "type": "integer",
          "format": "int64"
        },
        "Links": {
          "description": "Tested links",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Link"
          }
        }
      },
      "x-go-name": "Response",
      "x-go-package": "github.com/axllent/mailsandbox/internal/linkcheck"
    },
...
    "Rule": {
      "description": "Rule struct",
      "type": "object",
      "properties": {
        "Description": {
          "description": "SpamAssassin rule description",
          "type": "string"
        },
        "Name": {
          "description": "SpamAssassin rule name",
          "type": "string"
        },
        "Score": {
          "description": "Spam rule score",
          "type": "number",
          "format": "double"
        }
      },
      "x-go-package": "github.com/axllent/mailsandbox/internal/spamassassin"
    },
    "SpamAssassinResponse": {
      "description": "Result is a SpamAssassin result",
      "type": "object",
      "properties": {
        "Error": {
          "description": "If populated will return an error string",
          "type": "string"
        },
        "IsSpam": {
          "description": "Whether the message is spam or not",
          "type": "boolean"
        },
        "Rules": {
          "description": "Spam rules triggered",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Rule"
          }
        },
        "Score": {
          "description": "Total spam score based on triggered rules",
          "type": "number",
          "format": "double"
        }
      },
      "x-go-name": "Result",
      "x-go-package": "github.com/axllent/mailsandbox/internal/spamassassin"
    }
  },
  "responses": {
...
    "ErrorResponse": {
      "description": "Server error will return with a 400 status code\nwith the error message in the body",
      "schema": {
        "type": "string"
      }
    },
...
    "JSONErrorResponse": {
      "description": "JSON error response",
      "schema": {
        "type": "object",
        "properties": {
          "Error": {
            "description": "Error message",
            "type": "string",
            "example": "invalid format"
          }
        }
      }
    },
...
    "NotFoundResponse": {
      "description": "Not found error will return a 404 status code",
      "schema": {
        "type": "string"
      }
    },
...
    "SendMessageResponse": {
      "description": "Confirmation message for HTTP send API",
      "schema": {
        "type": "object",
        "properties": {
          "ID": {
            "description": "Database ID",
            "type": "string",
            "example": "iAfZVVe2UQfNSG5BAjgYwa"
          }
        }
      }
    },
...
  }
}